refactor chat api modules
