+++ lib/api/chat/questions.js
@@ -1,0 +1,29 @@
+export const hybridOfferQuestions = [
+  { key: 'offerDescription', question: "What's your core product or service?", description: 'Core product or service' },
+  { key: 'targetAudience', question: 'Who is your target audience?', description: 'Target audience details' },
+  { key: 'painPoints', question: 'What pain points do they face?', description: 'Customer pain points' },
+  { key: 'solution', question: 'How do you solve these problems?', description: 'Solution approach' },
+  { key: 'pricing', question: "What's your pricing structure?", description: 'Pricing information' },
+  { key: 'clientResult', question: "What's a specific, real-world result you've helped a client achieve?", description: 'Specific client success story' }
+];
+
+export const workshopQuestions = [
+  { key: 'participantOutcomes', question: 'What specific outcomes or goals will participants achieve by the end of your workshop?', description: 'Participant outcomes and goals' },
+  { key: 'targetAudience', question: 'Who is your ideal workshop participant? Please describe their demographics, current situation, and main pain points.', description: 'Target audience demographics and pain points' },
+  { key: 'problemAddressed', question: 'What specific problem or challenge does your workshop solve for these participants?', description: 'Problem the workshop addresses' },
+  { key: 'workshopDuration', question: 'How long will your workshop be? Please specify the duration and format.', description: 'Workshop duration and format' },
+  { key: 'topicsAndActivities', question: 'What key topics will you cover and what activities will participants engage in during the workshop?', description: 'Topics covered and activities' },
+  { key: 'resourcesProvided', question: 'What resources, materials, or follow-up support will participants receive?', description: 'Resources and materials provided' }
+];
+
+export function calculateQuestionsAnswered(collectedAnswers, tool = 'hybrid-offer') {
+  if (!collectedAnswers) return 0;
+  const questionsArray = tool === 'workshop-generator' ? workshopQuestions : hybridOfferQuestions;
+  let count = 0;
+  for (const q of questionsArray) {
+    if (collectedAnswers[q.key] && collectedAnswers[q.key].trim().length > 0) {
+      count++;
+    }
+  }
+  return count;
+}
